+++ admin/livesuche.php
@@ -3,7 +3,7 @@
  * @copyright (c) JTL-Software-GmbH
  * @license http://jtl-url.de/jtlshoplicense
  */
-require_once dirname(__FILE__) . '/includes/admininclude.php';
+require_once __DIR__ . '/includes/admininclude.php';
 
 $oAccount->permission('MODULE_LIVESEARCH_VIEW', true, true);
 /** @global JTLSmarty $smarty */
@@ -13,7 +13,7 @@ setzeSprache();
 
 $hinweis           = '';
 $fehler            = '';
-$settingsIDs       = array(423, 425, 422, 437, 438);
+$settingsIDs       = [423, 425, 422, 437, 438];
 
 // Tabs
 if (strlen(verifyGPDataString('tab')) > 0) {
@@ -21,9 +21,7 @@ if (strlen(verifyGPDataString('tab')) > 0) {
 }
 
 // Suchanfrage Suche
-if (!isset($cLivesucheSQL)) {
-    $cLivesucheSQL = new stdClass();
-}
+$cLivesucheSQL = new stdClass();
 $cLivesucheSQL->cWhere = '';
 $cLivesucheSQL->cOrder = ' tsuchanfrage.nAnzahlGesuche DESC ';
 if (strlen(verifyGPDataString('cSuche')) > 0) {
@@ -71,12 +69,12 @@ if (verifyGPCDataInteger('nSort') > 0) {
     $smarty->assign('nSort', -1);
 }
 
-if (isset($_POST['livesuche']) && intval($_POST['livesuche']) === 1) { //Formular wurde abgeschickt
+if (isset($_POST['livesuche']) && (int)$_POST['livesuche'] === 1) { //Formular wurde abgeschickt
     // Suchanfragen aktualisieren
     if (isset($_POST['suchanfragenUpdate'])) {
         if (is_array($_POST['kSuchanfrageAll']) && count($_POST['kSuchanfrageAll']) > 0) {
             foreach ($_POST['kSuchanfrageAll'] as $kSuchanfrage) {
-                if (strlen($_POST['nAnzahlGesuche_' . $kSuchanfrage]) > 0 && intval($_POST['nAnzahlGesuche_' . $kSuchanfrage]) > 0) {
+                if (strlen($_POST['nAnzahlGesuche_' . $kSuchanfrage]) > 0 && (int)$_POST['nAnzahlGesuche_' . $kSuchanfrage] > 0) {
                     $_upd                 = new stdClass();
                     $_upd->nAnzahlGesuche = (int)$_POST['nAnzahlGesuche_' . $kSuchanfrage];
                     Shop::DB()->update('tsuchanfrage', 'kSuchanfrage', (int)$kSuchanfrage, $_upd);
@@ -120,7 +118,7 @@ if (isset($_POST['livesuche']) && intval($_POST['livesuche']) === 1) { //Formula
             if (isset($_POST['nAktiv']) && is_array($_POST['nAktiv'])) {
                 foreach ($_POST['nAktiv'] as $i => $nAktiv) {
                     $oSuchanfrage = Shop::DB()->select('tsuchanfrage', 'kSuchanfrage', (int)$nAktiv);
-                    Shop::DB()->delete('tseo', array('cKey', 'kKey', 'kSprache'), array('kSuchanfrage', (int)$nAktiv, (int)$_SESSION['kSprache']));
+                    Shop::DB()->delete('tseo', ['cKey', 'kKey', 'kSprache'], ['kSuchanfrage', (int)$nAktiv, (int)$_SESSION['kSprache']]);
                     // Aktivierte Suchanfragen in tseo eintragen
                     $oSeo           = new stdClass();
                     $oSeo->cSeo     = checkSeo(getSeo($oSuchanfrage->cSuche));
@@ -139,33 +137,36 @@ if (isset($_POST['livesuche']) && intval($_POST['livesuche']) === 1) { //Formula
 
         if (count($Suchanfragen) > 0) {
             foreach ($Suchanfragen as $sucheanfrage) {
-                if (strtolower($sucheanfrage->cSuche) != strtolower($_POST['mapping_' . $sucheanfrage->kSuchanfrage])) {
-                    if (strlen($_POST['mapping_' . $sucheanfrage->kSuchanfrage]) > 0) {
+                if (!isset($_POST['mapping_' . $sucheanfrage->kSuchanfrage]) ||
+                    strtolower($sucheanfrage->cSuche) !== strtolower($_POST['mapping_' . $sucheanfrage->kSuchanfrage])) {
+                    if (!empty($_POST['mapping_' . $sucheanfrage->kSuchanfrage])) {
                         $nMappingVorhanden                      = 1;
                         $suchanfragemapping_obj                 = new stdClass();
                         $suchanfragemapping_obj->kSprache       = $_SESSION['kSprache'];
                         $suchanfragemapping_obj->cSuche         = $sucheanfrage->cSuche;
                         $suchanfragemapping_obj->cSucheNeu      = $_POST['mapping_' . $sucheanfrage->kSuchanfrage];
                         $suchanfragemapping_obj->nAnzahlGesuche = $sucheanfrage->nAnzahlGesuche;
-
                         $Neuesuche = Shop::DB()->select('tsuchanfrage', 'cSuche', $suchanfragemapping_obj->cSucheNeu);
-                        if ($Neuesuche->kSuchanfrage > 0) {
+                        if (isset($Neuesuche->kSuchanfrage) && $Neuesuche->kSuchanfrage > 0) {
                             Shop::DB()->insert('tsuchanfragemapping', $suchanfragemapping_obj);
                             Shop::DB()->query(
                                 "UPDATE tsuchanfrage
                                     SET nAnzahlGesuche = nAnzahlGesuche+" . $sucheanfrage->nAnzahlGesuche . "
                                     WHERE kSprache = " . (int)$_SESSION['kSprache'] . "
-                                        AND cSuche = '" . Shop::DB()->escape($_POST['mapping_' . $sucheanfrage->kSuchanfrage]) . "'", 4);
+                                        AND cSuche = '" . Shop::DB()->escape($_POST['mapping_' . $sucheanfrage->kSuchanfrage]) . "'", 4
+                            );
                             Shop::DB()->delete('tsuchanfrage', 'kSuchanfrage', (int)$sucheanfrage->kSuchanfrage);
                             $upd       = new stdClass();
                             $upd->kKey = (int)$Neuesuche->kSuchanfrage;
                             Shop::DB()->update('tseo', ['cKey', 'kKey'], ['kSuchanfrage', (int)$sucheanfrage->kSuchanfrage], $upd);
 
-                            $hinweis .= 'Die Suchanfrage "' . $suchanfragemapping_obj->cSuche . '" wurde erfolgreich auf "' . $suchanfragemapping_obj->cSucheNeu . '" gemappt.<br />';
+                            $hinweis .= 'Die Suchanfrage "' . $suchanfragemapping_obj->cSuche .
+                                '" wurde erfolgreich auf "' . $suchanfragemapping_obj->cSucheNeu . '" gemappt.<br />';
                         }
                     }
                 } else {
-                    $fehler .= 'Die Suchanfrage "' . $sucheanfrage->cSuche . '" kann nicht auf den gleichen Suchebegriff gemappt werden.';
+                    $fehler .= 'Die Suchanfrage "' . $sucheanfrage->cSuche .
+                        '" kann nicht auf den gleichen Suchebegriff gemappt werden.';
                 }
             }
         }
@@ -180,7 +181,7 @@ if (isset($_POST['livesuche']) && intval($_POST['livesuche']) === 1) { //Formula
                     $oSuchanfrage = Shop::DB()->select('tsuchanfrage', 'kSuchanfrage', (int)$kSuchanfrage);
 
                     if ($oSuchanfrage->kSuchanfrage > 0) {
-                        if (strtolower($oSuchanfrage->cSuche) != strtolower($cMapping)) {
+                        if (strtolower($oSuchanfrage->cSuche) !== strtolower($cMapping)) {
                             $oSuchanfrageNeu = Shop::DB()->select('tsuchanfrage', 'cSuche', Shop::DB()->escape($cMapping));
                             if (isset($oSuchanfrageNeu->kSuchanfrage) && $oSuchanfrageNeu->kSuchanfrage > 0) {
                                 $oSuchanfrageMapping                 = new stdClass();
@@ -238,7 +239,7 @@ if (isset($_POST['livesuche']) && intval($_POST['livesuche']) === 1) { //Formula
                 Shop::DB()->delete('tsuchanfrage', 'kSuchanfrage', (int)$kSuchanfrage);
                 Shop::DB()->insert('tsuchanfrageblacklist', $obj);
                 // Aus tseo loeschen
-                Shop::DB()->delete('tseo', array('cKey', 'kKey'), array('kSuchanfrage', (int)$kSuchanfrage));
+                Shop::DB()->delete('tseo', ['cKey', 'kKey'], ['kSuchanfrage', (int)$kSuchanfrage]);
                 $hinweis .= 'Die Suchanfrage "' . $kSuchanfrage_obj->cSuche . '" wurde erfolgreich gel&ouml;scht.<br />';
                 $hinweis .= 'Die Suchanfrage "' . $kSuchanfrage_obj->cSuche . '" wurde auf die Blacklist hinzugef&uuml;gt.<br />';
             }
@@ -246,14 +247,21 @@ if (isset($_POST['livesuche']) && intval($_POST['livesuche']) === 1) { //Formula
             $fehler .= 'Bitte w&auml;hlen Sie mindestens eine Suchanfrage aus.<br />';
         }
     }
-} elseif (isset($_POST['livesuche']) && intval($_POST['livesuche']) === 2) { // Erfolglos mapping
+} elseif (isset($_POST['livesuche']) && (int)$_POST['livesuche'] === 2) { // Erfolglos mapping
     if (isset($_POST['erfolglosEdit'])) { // Editieren
         $smarty->assign('nErfolglosEditieren', 1);
     } elseif (isset($_POST['erfolglosUpdate'])) { // Update
-        $Suchanfragenerfolglos = Shop::DB()->selectAll('tsuchanfrageerfolglos', 'kSprache', (int)$_SESSION['kSprache'], '*', 'nAnzahlGesuche DESC');
+        $Suchanfragenerfolglos = Shop::DB()->selectAll(
+            'tsuchanfrageerfolglos',
+            'kSprache',
+            (int)$_SESSION['kSprache'],
+            '*',
+            'nAnzahlGesuche DESC'
+        );
         if (count($Suchanfragenerfolglos) > 0) {
             foreach ($Suchanfragenerfolglos as $Suchanfrageerfolglos) {
-                if (isset($_POST['mapping_' . $Suchanfrageerfolglos->kSuchanfrageErfolglos]) && strlen($_POST['mapping_' . $Suchanfrageerfolglos->kSuchanfrageErfolglos]) > 0) {
+                if (isset($_POST['mapping_' . $Suchanfrageerfolglos->kSuchanfrageErfolglos]) &&
+                    strlen($_POST['mapping_' . $Suchanfrageerfolglos->kSuchanfrageErfolglos]) > 0) {
                     if (strtolower($Suchanfrageerfolglos->cSuche) != strtolower($_POST['mapping_' . $Suchanfrageerfolglos->kSuchanfrageErfolglos])) {
                         if (strlen($_POST['mapping_' . $Suchanfrageerfolglos->kSuchanfrageErfolglos]) > 0) {
                             $suchanfragemapping_obj                 = new stdClass();
@@ -267,13 +275,15 @@ if (isset($_POST['livesuche']) && intval($_POST['livesuche']) === 1) { //Formula
                                 Shop::DB()->insert('tsuchanfragemapping', $suchanfragemapping_obj);
                                 Shop::DB()->delete('tsuchanfrageerfolglos', 'kSuchanfrageErfolglos', (int)$oAlteSuche->kSuchanfrageErfolglos);
 
-                                $hinweis .= 'Die Suchanfrage "' . $suchanfragemapping_obj->cSuche . '" wurde erfolgreich auf "' . $suchanfragemapping_obj->cSucheNeu . '" gemappt.<br />';
+                                $hinweis .= 'Die Suchanfrage "' . $suchanfragemapping_obj->cSuche .
+                                    '" wurde erfolgreich auf "' . $suchanfragemapping_obj->cSucheNeu . '" gemappt.<br />';
                             }
                         }
                     } else {
-                        $fehler .= 'Die Suchanfrage "' . $Suchanfrageerfolglos->cSuche . '" kann nicht auf den gleichen Suchbegriff gemappt werden.';
+                        $fehler .= 'Die Suchanfrage "' . $Suchanfrageerfolglos->cSuche .
+                            '" kann nicht auf den gleichen Suchbegriff gemappt werden.';
                     }
-                } elseif (intval($_POST['nErfolglosEditieren']) === 1) {
+                } elseif ((int)$_POST['nErfolglosEditieren'] === 1) {
                     $Suchanfrageerfolglos->cSuche = StringHandler::filterXSS($_POST['cSuche_' . $Suchanfrageerfolglos->kSuchanfrageErfolglos]);
                     $upd                          = new stdClass();
                     $upd->cSuche                  = $Suchanfrageerfolglos->cSuche;
@@ -294,7 +304,7 @@ if (isset($_POST['livesuche']) && intval($_POST['livesuche']) === 1) { //Formula
         }
     }
     $smarty->assign('tab', 'erfolglos');
-} elseif (isset($_POST['livesuche']) && intval($_POST['livesuche']) === 3) { // Blacklist
+} elseif (isset($_POST['livesuche']) && (int)$_POST['livesuche'] === 3) { // Blacklist
     $suchanfragenblacklist = $_POST['suchanfrageblacklist'];
     $suchanfragenblacklist = explode(';', $suchanfragenblacklist);
     $count                 = count($suchanfragenblacklist);
@@ -310,7 +320,7 @@ if (isset($_POST['livesuche']) && intval($_POST['livesuche']) === 1) { //Formula
     }
     $smarty->assign('tab', 'blacklist');
     $hinweis .= 'Die Blacklist wurde erfolgreich aktualisiert.';
-} elseif (isset($_POST['livesuche']) && intval($_POST['livesuche']) === 4) { // Mappinglist
+} elseif (isset($_POST['livesuche']) && (int)$_POST['livesuche'] === 4) { // Mappinglist
     if (isset($_POST['delete'])) {
         if (is_array($_POST['kSuchanfrageMapping'])) {
             foreach ($_POST['kSuchanfrageMapping'] as $kSuchanfrageMapping) {
@@ -421,7 +431,7 @@ for ($i = 0; $i < $configCount; $i++) {
                 AND cName = '" . $oConfig_arr[$i]->cWertName . "'", 1
     );
 
-    $oConfig_arr[$i]->gesetzterWert = (isset($oSetValue->cWert)) ? $oSetValue->cWert : null;
+    $oConfig_arr[$i]->gesetzterWert = isset($oSetValue->cWert) ? $oSetValue->cWert : null;
 }
 $smarty->assign('oConfig_arr', $oConfig_arr)
     ->assign('Sprachen', $Sprachen)
