+++ classes/class.JTL-Shop.Link.php
@@ -67,19 +67,19 @@ class Link extends MainModel
     public $cDruckButton;
 
     /**
-     * @var string
+     * @var int
      */
     public $nSort;
 
     /**
-     * @var string
+     * @var int
      */
-    public $bSSL = '0';
+    public $bSSL = 0;
 
     /**
-     * @var string
+     * @var int
      */
-    public $bIsFluid = '0';
+    public $bIsFluid = 0;
 
     /**
      * @var string
@@ -94,7 +94,185 @@ class Link extends MainModel
     /**
      * @var array
      */
-    public $oSub_arr = array();
+    public $oSub_arr = [];
+
+    /**
+     * @var string
+     */
+    public $cISO;
+
+    /**
+     * @var int
+     */
+    public $kSprache = 0;
+
+    /**
+     * @var string
+     */
+    public $cSeo;
+
+    /**
+     * @var int
+     */
+    public $nHTTPRedirectCode = 0;
+
+    /**
+     * @var bool
+     */
+    public $bHideContent = false;
+
+    /**
+     * @var int
+     */
+    public $nPluginStatus = 0;
+
+    /**
+     * @var string
+     */
+    public $cURLFull;
+
+    /**
+     * @var string
+     */
+    public $cURLFullSSL;
+
+    /**
+     * @var int
+     */
+    public $kSpezialSeite = 0;
+
+    /**
+     * @param int $kSpezialSeite
+     * @return $this
+     */
+    public function setSpezialSeite($kSpezialSeite)
+    {
+        $this->kSpezialSeite = (int)$kSpezialSeite;
+
+        return $this;
+    }
+
+    /**
+     * @return int
+     */
+    public function getSpezialSeite()
+    {
+        return (int)$this->kSpezialSeite;
+    }
+
+    /**
+     * @return string
+     */
+    public function getURLFullSSL()
+    {
+        return $this->cURLFullSSL;
+    }
+
+    /**
+     * @param string $cURLFullSSL
+     * @return $this
+     */
+    public function setURLFullSSL($cURLFullSSL)
+    {
+        $this->cURLFullSSL = $cURLFullSSL;
+
+        return $this;
+    }
+
+    /**
+     * @return string
+     */
+    public function getURLFull()
+    {
+        return $this->cURLFull;
+    }
+
+    /**
+     * @param string $cURLFull
+     * @return $this
+     */
+    public function setURLFull($cURLFull)
+    {
+        $this->cURLFull = $cURLFull;
+
+        return $this;
+    }
+
+    /**
+     * @param int $nPluginStatus
+     * @return $this
+     */
+    public function setPluginStatus($nPluginStatus)
+    {
+        $this->nPluginStatus = (int)$nPluginStatus;
+
+        return $this;
+    }
+
+    /**
+     * @return int
+     */
+    public function getPluginStatus()
+    {
+        return (int)$this->nPluginStatus;
+    }
+
+    /**
+     * @param string $cISO
+     * @return $this
+     */
+    public function setISO($cISO)
+    {
+        $this->cISO = $cISO;
+
+        return $this;
+    }
+
+    /**
+     * @return string
+     */
+    public function getISO()
+    {
+        return $this->cISO;
+    }
+
+    /**
+     * @param int $kSprache
+     * @return $this
+     */
+    public function setSprache($kSprache)
+    {
+        $this->kSprache = (int)$kSprache;
+
+        return $this;
+    }
+
+    /**
+     * @return int
+     */
+    public function getSprache()
+    {
+        return (int)$this->kSprache;
+    }
+
+    /**
+     * @param string $cSeo
+     * @return $this
+     */
+    public function setSeo($cSeo)
+    {
+        $this->cSeo = $cSeo;
+
+        return $this;
+    }
+
+    /**
+     * @return string
+     */
+    public function getSeo()
+    {
+        return $this->cSeo;
+    }
 
     /**
      * @return int
@@ -314,7 +492,7 @@ class Link extends MainModel
      */
     public function getSort()
     {
-        return $this->nSort;
+        return (int)$this->nSort;
     }
 
     /**
@@ -329,12 +507,12 @@ class Link extends MainModel
     }
 
     /**
-     * @param bool $mode
+     * @param int $mode
      * @return $this
      */
     public function setSSL($mode)
     {
-        $this->bSSL = $mode;
+        $this->bSSL = (int)$mode;
 
         return $this;
     }
@@ -344,26 +522,26 @@ class Link extends MainModel
      */
     public function getSSL()
     {
-        return $this->bSSL;
+        return (int)$this->bSSL;
     }
 
     /**
-     * @param $mode
+     * @param int $mode
      * @return $this
      */
     public function setIsFluid($mode)
     {
-        $this->bIsFluid = $mode;
+        $this->bIsFluid = (int)$mode;
 
         return $this;
     }
 
     /**
-     * @return string
+     * @return int
      */
     public function getIsFluid()
     {
-        return $this->bIsFluid;
+        return (int)$this->bIsFluid;
     }
 
     /**
@@ -386,20 +564,39 @@ class Link extends MainModel
     }
 
     /**
+     * @param null|int    $kKey
+     * @param null|object $oObj
+     * @param null|mixed  $xOption
+     * @param nulll|int   $kLinkgruppe
+     */
+    public function __construct($kKey = null, $oObj = null, $xOption = null, $kLinkgruppe = null)
+    {
+        if (is_object($oObj)) {
+            $this->loadObject($oObj);
+        } elseif ($kKey !== null) {
+            $this->load($kKey, $oObj, $xOption, $kLinkgruppe);
+        }
+    }
+
+    /**
      * @param int         $kKey
      * @param object|null $oObj
      * @param mixed|null  $xOption
+     * @param int         $kLinkgruppe
      * @return $this
      */
-    public function load($kKey, $oObj = null, $xOption = null)
+    public function load($kKey, $oObj = null, $xOption = null, $kLinkgruppe = null)
     {
-        $oObj = Shop::DB()->select('tlink', 'kLink', (int)$kKey);
-
+        if (!empty($kLinkgruppe)) {
+            $oObj = Shop::DB()->select('tlink', ['kLink', 'kLinkgruppe'], [(int)$kKey, (int)$kLinkgruppe]);
+        } else {
+            $oObj = Shop::DB()->select('tlink', 'kLink', (int)$kKey);
+        }
         if (!empty($oObj->kLink)) {
             $this->loadObject($oObj);
 
             if ($xOption) {
-                $this->oSub_arr = self::getSub($this->getLink());
+                $this->oSub_arr = self::getSub($this->getLink(), $this->getLinkgruppe());
             }
         }
 
@@ -408,19 +605,24 @@ class Link extends MainModel
 
     /**
      * @param int $kVaterLink
+     * @param int $kVaterLinkgruppe
      * @return null|array
      */
-    public static function getSub($kVaterLink)
+    public static function getSub($kVaterLink, $kVaterLinkgruppe = null)
     {
-        $kVaterLink = (int)$kVaterLink;
+        $kVaterLink       = (int)$kVaterLink;
+        $kVaterLinkgruppe = (int)$kVaterLinkgruppe;
         if ($kVaterLink > 0) {
-            $oLink_arr = Shop::DB()->selectAll('tlink', 'kVaterLink', $kVaterLink);
+            if (!empty($kVaterLinkgruppe)) {
+                $oLink_arr = Shop::DB()->selectAll('tlink', ['kVaterLink', 'kLinkgruppe'],
+                    [$kVaterLink, $kVaterLinkgruppe]);
+            } else {
+                $oLink_arr = Shop::DB()->selectAll('tlink', 'kVaterLink', $kVaterLink);
+            }
 
             if (is_array($oLink_arr) && count($oLink_arr) > 0) {
                 foreach ($oLink_arr as &$oLink) {
-                    $kLinkgruppe        = $oLink->kLinkgruppe;
-                    $oLink              = new self($oLink->kLink, null, true);
-                    $oLink->kLinkgruppe = (int)$kLinkgruppe;
+                    $oLink = new self($oLink->kLink, null, true, $kVaterLinkgruppe);
                 }
             }
 
@@ -473,7 +675,7 @@ class Link extends MainModel
     public function update()
     {
         $cQuery   = "UPDATE tlink SET ";
-        $cSet_arr = array();
+        $cSet_arr = [];
 
         $cMember_arr = array_keys(get_object_vars($this));
         if (is_array($cMember_arr) && count($cMember_arr) > 0) {
@@ -482,17 +684,17 @@ class Link extends MainModel
                 if (method_exists($this, $cMethod)) {
                     $mValue = "'" . Shop::DB()->realEscape(
                             call_user_func(
-                                array(
+                                [
                                     &$this,
                                     $cMethod
-                                )
+                                ]
                             )
                         ) . "'";
                     if (call_user_func(
-                            array(
+                            [
                                 &$this,
                                 $cMethod
-                            )
+                            ]
                         ) === null
                     ) {
                         $mValue = 'NULL';
@@ -502,7 +704,7 @@ class Link extends MainModel
             }
 
             $cQuery .= implode(', ', $cSet_arr);
-            $cQuery .= " WHERE kLink = {$this->getLink()}";
+            $cQuery .= " WHERE kLink = {$this->getLink()} AND klinkgruppe = {$this->getLinkgruppe()}";
 
             return Shop::DB()->query($cQuery, 3);
         } else {
@@ -530,18 +732,14 @@ class Link extends MainModel
                 Shop::DB()->delete('tseo', ['kKey', 'cKey'], [$this->getLink(), 'kLink']);
 
                 $cDir = PFAD_ROOT . PFAD_BILDER . PFAD_LINKBILDER . $this->getLink();
-                if (is_dir($cDir) && $this->getLink() > 0) {
-                    if (delDirRecursively($cDir)) {
-                        rmdir($cDir);
-                    }
+                if (is_dir($cDir) && $this->getLink() > 0 && delDirRecursively($cDir)) {
+                    rmdir($cDir);
                 }
             }
 
-            if ($bSub) {
-                if (isset($this->oSub_arr) && count($this->oSub_arr) > 0) {
-                    foreach ($this->oSub_arr as $oSub) {
-                        $oSub->delete(true, $kLinkgruppe);
-                    }
+            if ($bSub && count($this->oSub_arr) > 0) {
+                foreach ($this->oSub_arr as $oSub) {
+                    $oSub->delete(true, $kLinkgruppe);
                 }
             }
         }
